Docker/Web/requirements.txt
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.8.2
numpy==1.15.2
pandas==0.23.4
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
scikit-learn==0.20.0
mecab-python3==0.7
